Rename wagtail.utils.apps to wagtail.utils.app
