+++ wagtail/images/formats.py
@@ -1,7 +1,7 @@
 from django.utils.html import escape
 from django.utils.translation import gettext_lazy as _
 
-from wagtail.utils.apps import get_app_submodules
+from wagtail.utils.app import get_app_submodules
 
 from .shortcuts import get_rendition_or_not_found
